requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
dj-rest-auth==5.1.0
Django==4.2.11
django-allauth==0.61.1
django-filter==24.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
h11==0.14.0
idna==3.6
Markdown==3.6
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.2.1
uvicorn==0.29.0
whitenoise==6.6.0
